requirements.txt
python-telegram-bot==20.7
flask
pymongo
dnspython
python-dotenv
